requirements.txt
pymongo
flask_pymongo
gitpython
requests